+++ packages/orq-rc/RELEASES.md
@@ -8429,3 +8429,13 @@ Based on:
 - [typescript v4.8.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.35 - packages/orq-rc
+
+## 2026-04-21 09:37:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.36 - packages/orq-rc